% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsLegacy-env.R
\name{getEsqlabsLegacyRSetting}
\alias{getEsqlabsLegacyRSetting}
\title{Get the value of a global esqlabsRLegacy setting.}
\usage{
getEsqlabsLegacyRSetting(settingName)
}
\arguments{
\item{settingName}{String name of the setting}
}
\value{
Value of the setting stored in esqlabsLegacyEnv. If the setting does not
exist, an error is thrown.
}
\description{
Get the value of a global esqlabsRLegacy setting.
}
\examples{
getEsqlabsRSetting("packageVersion")
getEsqlabsRSetting("widthPerPlotMapping")
}
